Added initial support for OpenSUSE Leap

+++ template_scripts/distribution.sh
@@ -41,6 +41,17 @@ if [ "${DIST#centos}" != "${DIST}" ]; then
     fi
 fi
 
+if [ "${DIST#lp}" != "${DIST}" ]; then
+    DISTRIBUTION="opensuse"
+	DIST_VER="${DIST#lp}"
+	DIST_VER=$(sed 's/.\{1\}$/.&/' <<< "$DIST_VER")
+
+    if [ -n "${FEDORA_MIRROR}" ]; then
+        YUM_OPTS="$YUM_OPTS --setopt=oss.baseurl=${OPENSUSE_MIRROR%/}/distribution/leap/${DIST_VER}/repo/oss/x86_64"
+        YUM_OPTS="$YUM_OPTS --setopt=updates.baseurl=${OPENSUSE_MIRROR%/}/update/leap/${DIST_VER}/oss/x86_64"
+    fi
+fi
+
 # ==============================================================================
 # Cleanup function
 # ==============================================================================
@@ -88,7 +99,7 @@ function yumInstall() {
     if [ -e "${INSTALLDIR}/usr/bin/$YUM" ]; then
         cp ${SCRIPTSDIR}/template-builder-repo-$DISTRIBUTION.repo ${INSTALLDIR}/etc/yum.repos.d/
         chroot_cmd $YUM --downloadonly \
-            install ${YUM_OPTS} -y ${files[@]} || exit 1
+            install --allowerasing ${YUM_OPTS} -v -y ${files[@]} || exit 1
         find ${INSTALLDIR}/var/cache/dnf -name '*.rpm' -print0 | xargs -r0 sha256sum
         find ${INSTALLDIR}/var/cache/yum -name '*.rpm' -print0 | xargs -r0 sha256sum
         if [ "$DISTRIBUTION" = "fedora" ]; then
@@ -101,6 +112,10 @@ function yumInstall() {
         if [ "$DISTRIBUTION" = "centos" ]; then
             # Temporarly disable previous strategy (problem with downloading cache qubes template repo)
             chroot_cmd $YUM install ${YUM_OPTS} -y ${files[@]} || exit 1
+        fi
+        if [ "$DISTRIBUTION" = "opensuse" ]; then
+            # TODO Firewalld package causes problems without --allowerasing
+            chroot_cmd $YUM install --allowerasing ${YUM_OPTS} -y ${files[@]} || exit 1
         fi
         rm -f ${INSTALLDIR}/etc/yum.repos.d/template-builder-repo-$DISTRIBUTION.repo
     else
